[CONTINUUM-2015] added confirmation page when deleting purge configuration
Submitted By: Gwen Harold Autencio


git-svn-id: https://svn.eu.apache.org/repos/asf/continuum/trunk@739894 13f79535-47bb-0310-9956-ffa450edef68

+++ continuum-webapp/src/main/java/org/apache/continuum/web/action/admin/PurgeConfigurationAction.java
@@ -265,7 +265,14 @@ public class PurgeConfigurationAction
     public String remove()
         throws Exception
     {
-        purgeConfigService.removePurgeConfiguration( purgeConfigId );
+        if ( confirmed )
+        {
+        	purgeConfigService.removePurgeConfiguration( purgeConfigId );
+        }
+        else
+        {
+            return CONFIRM;
+        }
         
         return SUCCESS;
     }
